수정

+++ src/main/java/Code04_03.java
@@ -6,6 +6,6 @@ public class Code04_03 {
         System.out.println(n1+n2);
         System.out.println(n1-n2);
         System.out.println(n1*n2);
-        System.out.println(n1/n2);
+        System.out.println((double)n1/n2);
     }
 }
